pyproject.toml
[project]
name = "cv-rank-agent"
version = "0.1.0"
description = "A Python agent that ranks a CV against job opening URLs using LLM + cosine similarity."
readme = "README.md"
authors = [
    { name = "Victor Manoliu", email = "victor_manoliu@epam.com" }
]
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "crawl4ai>=0.8.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.9",
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
]

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"
